Updating from BIC (or whatever) to likelihood
Some notes: it turns out that because the likelihood is cross-validated here we can basically ignore the 2p and log(n)p terms in AIC/BIC and consider the delta likelihood as equivalent to these. This also means that exp(delta cvLL) is the likelihood ratio, so the difference in cross-validated likelihood is directly intepretable.